Changed $M::MIN into $M::MIN_POSITIVE in the impl_real! macro.
The docs for the function state "The smallest finite positive value representable using this type.", but f32::MIN & f64::MIN are both the smallest values representable(and are thus negative). It appears impl_real! is only ever used for f32 and f64, both of which have the "::MIN_VALUE constant. This change makes the functionality of the ::min_value() method consistent with the docs.