feat: update `rand` to 0.9
Implementations of `Distribution<T> for StandardUniform` for all
SIMD types containing `isize` and `usize` now generate random
values using `fill_bytes()`.

Uniform distribution for `isize` and `usize` was removed from `rand` crate.
See https://github.com/rust-random/rand/pull/1487 for details.

There is a new `UniformUsize` distribution, but it is limited to a
maximum value under u32::MAX for portability across 32/64bits and for
use as array indexes and lengths. Probably it is not suitable if a real
uniform distribution is required.